+++ CHANGES.rst
@@ -10,6 +10,10 @@
   <https://github.com/NextThought/nti.fakestatsd/issues/6>`_.
 - Metric values are now always native strings. See `issue 2
   <https://github.com/NextThought/nti.fakestatsd/issues/2>`_.
+- Restructure and document public objects. See `issue 1
+  <https://github.com/NextThought/nti.fakestatsd/issues/1>`_.
+- Add convenience hamcrest matchers. See `issue 1
+  <https://github.com/NextThought/nti.fakestatsd/issues/1>`_.
 
 
 0.0.1 (2018-10-12)